Read sensor data project structure

+++ db/db.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS raw_sensor_data;
 DROP TABLE IF EXISTS node_metadata;
 DROP TABLE IF EXISTS sensor_metadata;
 DROP TABLE IF EXISTS node_sensor;
-DROP TABLE IF EXISTS network_metadata; 
+DROP TABLE IF EXISTS network_metadata;
 
 
 CREATE TABLE network_metadata (
@@ -30,20 +30,20 @@ CREATE TABLE node_sensor (
         PRIMARY KEY (
                 node_id,
                 sensor_id
-        ), 
+        ),
         FOREIGN KEY (node_id) REFERENCES node_metadata(node_id),
-        FOREIGN KEY (sensor_id) REFERENCES sensor_metadata (sensor_id) 
+        FOREIGN KEY (sensor_id) REFERENCES sensor_metadata (sensor_id)
 );
 
 CREATE TABLE raw_sensor_data (
         node_id INTEGER NOT NULL ,
         time DATETIME NOT NULL ,
         sensor_id INTEGER NOT NULL ,
-        value INTEGER NOT NULL , 
+        value DOUBLE NOT NULL , 
         PRIMARY KEY (
                 node_id,
                 time,
                 sensor_id
-        ), 
+        ),
         FOREIGN KEY (node_id, sensor_id) REFERENCES node_sensor (node_id, sensor_id),
 );
